+++ packages/template-retail-react-app/app/pages/product-list/partials/inventory-filter.jsx
@@ -91,6 +91,7 @@ const StoreInventoryFilter = ({toggleFilter, selectedFilters}) => {
                     />
                 </Heading>
                 <Checkbox
+                    data-testid="sf-store-inventory-filter-checkbox"
                     isChecked={isChecked}
                     onChange={handleCheckboxChange}
                     aria-label={formatMessage(