+++ packages/template-retail-react-app/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Compute store level inventory and refactor product view [#2553](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2553)
 - Show Bonus Product Label on OrderSummary component [#2524](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2524)
 - Updated store-locator component to work with Search Engines and location embedding [#2551](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2551)
+- Added additonal fixes for PDP page to support pickup in store[#2593] (https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2593)
 
 ## v6.1.0 (May 22, 2025)